+++ crates/bevy_image/src/exr_texture_loader.rs
@@ -1,5 +1,5 @@
-use crate::{Image, RenderAssetUsages, TextureFormatPixelInfo};
-use bevy_asset::{io::Reader, AssetLoader, LoadContext};
+use crate::{Image, TextureFormatPixelInfo};
+use bevy_asset::{io::Reader, AssetLoader, LoadContext, RenderAssetUsages};
 use image::ImageDecoder;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
