+++ crates/bevy_image/src/hdr_texture_loader.rs
@@ -1,4 +1,5 @@
-use crate::{Image, RenderAssetUsages, TextureFormatPixelInfo};
+use crate::{Image, TextureFormatPixelInfo};
+use bevy_asset::RenderAssetUsages;
 use bevy_asset::{io::Reader, AssetLoader, LoadContext};
 use image::DynamicImage;
 use serde::{Deserialize, Serialize};
