+++ crates/bevy_image/src/image_texture_conversion.rs
@@ -1,4 +1,5 @@
-use crate::{Image, RenderAssetUsages, TextureFormatPixelInfo};
+use crate::{Image, TextureFormatPixelInfo};
+use bevy_asset::RenderAssetUsages;
 use image::{DynamicImage, ImageBuffer};
 use thiserror::Error;
 use wgpu::{Extent3d, TextureDimension, TextureFormat};
