+++ crates/bevy_render/src/texture/fallback_image.rs
@@ -1,4 +1,5 @@
 use crate::{
+    render_asset::RenderAssetUsages,
     render_resource::*,
     renderer::{RenderDevice, RenderQueue},
     texture::{DefaultImageSampler, GpuImage},
@@ -8,7 +9,7 @@ use bevy_ecs::{
     prelude::{FromWorld, Res, ResMut},
     system::{Resource, SystemParam},
 };
-use bevy_image::{BevyDefault, Image, ImageSampler, RenderAssetUsages, TextureFormatPixelInfo};
+use bevy_image::{BevyDefault, Image, ImageSampler, TextureFormatPixelInfo};
 use bevy_utils::HashMap;
 
 /// A [`RenderApp`](crate::RenderApp) resource that contains the default "fallback image",
